+++ crates/ruff_linter/resources/test/fixtures/pylint/non_slot_assignment.py
