+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -1,6 +1,5 @@
 pub(crate) use and_or_ternary::*;
 pub(crate) use assert_on_string_literal::*;
-pub(crate) use assigning_non_slot::*;
 pub(crate) use await_outside_async::*;
 pub(crate) use bad_dunder_method_name::*;
 pub(crate) use bad_open_mode::*;
@@ -42,6 +41,7 @@ pub(crate) use no_method_decorator::*;
 pub(crate) use no_self_use::*;
 pub(crate) use non_ascii_module_import::*;
 pub(crate) use non_ascii_name::*;
+pub(crate) use non_slot_assignment::*;
 pub(crate) use nonlocal_without_binding::*;
 pub(crate) use potential_index_error::*;
 pub(crate) use property_with_parameters::*;
@@ -84,7 +84,6 @@ pub(crate) use yield_in_init::*;
 
 mod and_or_ternary;
 mod assert_on_string_literal;
-mod assigning_non_slot;
 mod await_outside_async;
 mod bad_dunder_method_name;
 mod bad_open_mode;
@@ -126,6 +125,7 @@ mod no_method_decorator;
 mod no_self_use;
 mod non_ascii_module_import;
 mod non_ascii_name;
+mod non_slot_assignment;
 mod nonlocal_without_binding;
 mod potential_index_error;
 mod property_with_parameters;
